Add ingress and egress nft logic for non zone policies
By choosing ingress and egress zones for a policy, it is wanted the
traffic to be forwarded from the ingress zones to the egress zones. It
is also wanted that any configured filtering for that policy is applied
to traffic forwarded between the ingress and egress zones.

This patch adds a new chain for each policy and a new path from the
'filter_FORWARD' chain to each policy chain created. The policy chain
will first check whether the packet applies to itself, and then continue
if it doesn't, or, filter the packet according to the policy rules.