test(policy): reload settings in test
In testing the setting of configuration for runtime settings, previous
configurations from testing permanent configuration would remain active.
This result in a failure saying that some configuration has already been
set.

./masquerade.at:36
+Warning: ALREADY_ENABLED: rule family="ipv4" source address="10.10.10.0/24" masquerade

This patch adds a reload after to refresh the configuration for the next
steps in the test. This is done for the masquerade test for a policy.