feat(policy): config check for runtime policy
When configuration for a policy is set into permanent settings it has a
set of checks done in the _check_config() function. These checks are not
applied to configuration for runtime settings.

This was failing test masquerade.at:68 which was intended to raise an
INVALID_ZONE error. The fix proposed creates a temporary policy object
to perform the checks.