feat(policy): check config based on future config
Ingress and egress zones are checked for validity in _check_config().
This function raises an error for incorrect config and checks for
incorrect config based on its current state. The assumption that all
future config can be compared with current config is incorrect for
config which relies on other config to be correct/incorrect, for
example, ingress and egress zones.

In policy.at:66 this flaw is exposed when attempting to change the
egress zone from 'HOST' to 'internal'. The future value, 'internal',
produced and error because the checks were being executed on the
assumption that 'HOST' would be the resulting egress zone.

This patch singles out ingress and egress configuration to be compared
with its future value.