fix(rich_rules): invalid nft log levels
Nftables has set values used to set the logging level for an nft rule
and these don't align exactly with the logging levels iptables uses.
'warn' and 'err' are expected from nftables where iptables (at least for
v1.8.4) expects 'warning' and 'error'. So when using nftables as the
backend, the resulting error was produced when using a log level of
'warning'.

	ERROR: COMMAND_FAILED: 'python-nftables' failed: internal:0:0-0: Error: Invalid log level 'warning'.

This fix converts the log level for nftables before it creates the
chain.