fix(tests): ignore expected errors in tests
The test suite tests for errors in the code, and the tests can xfail to
not produce a fail for the test. The test errors produce errors and
warnings in the logs which trigger a failure at the end of a test. These
logs must also be caught in the test so that the test doesn't result in
a fail.